QA Merge: W37/38/39 1st

+++ run_terminate_imagebuilder_gendo_old_environment.py
@@ -1,7 +1,11 @@
 #!/usr/bin/env python3
 
 import re
+import time
+from datetime import datetime
+from datetime import timedelta
 
+from env import env
 from run_common import AWSCli
 from run_common import print_message
 from run_common import print_session
@@ -63,7 +67,6 @@ def delete_version_list_any_ec2_image(delete_versions, image):
     for vv in delete_versions:
         for tag in image['Tags']:
             if tag['Key'] == 'Ec2ImageBuilderArn' and vv in tag['Value']:
-                print(tag['Value'])
                 return True
     return False
 
@@ -99,32 +102,68 @@ def run_terminate_image(name):
     account_id = aws_cli.get_caller_account_id()
     imagebuilder_resource = dict()
 
-    # print_message(f'get component list')
-    cmd = ['imagebuilder', 'list-components']
-    rr = aws_cli.run(cmd)
+    print_message('check in used ami version')
+    ec2_describe_role_arn = ''
+    for settings in env.get('imagebuilder', list()):
+        if settings['NAME'] == 'run_terminate_imagebuilder_gendo_old_environment':
+            ec2_describe_role_arn = settings['EC2_DESCRIBE_ROLE_ARN']
 
-    component_list = filter_imagebuilder_resource_arn_list(rr['componentVersionList'])
+    if ec2_describe_role_arn:
+        cmd = ['sts', 'assume-role']
+        cmd += ['--role-arn', ec2_describe_role_arn]
+        cmd += ['--role-session-name', 'ec2-describe-role']
+        rr = aws_cli.run(cmd)
 
-    timestamp_list = list()
-    for cc in component_list:
-        try:
-            found = re.search('-component-(.+?)/', cc).group(1)
-            if found:
-                timestamp_list.append(found)
-        except AttributeError:
-            pass
-    timestamp_list = set(timestamp_list)
+        access_key = rr['Credentials']['AccessKeyId']
+        secret_key = rr['Credentials']['SecretAccessKey']
+        session_token = rr['Credentials']['SessionToken']
+
+        aws_cli_for_ec2 = AWSCli(aws_access_key=access_key,
+                                 aws_secret_access_key=secret_key,
+                                 aws_session_token=session_token)
+
+        cmd = ['elasticbeanstalk', 'describe-environments']
+        rr = aws_cli_for_ec2.run(cmd)
+
+        gendo_eb_name_list = list()
+        for r in rr['Environments']:
+            if r['EnvironmentName'].startswith('gendo'):
+                gendo_eb_name_list.append((r['EnvironmentName']))
+
+        in_use_ec2_ami_list = list()
+        for eb_name in gendo_eb_name_list:
+            cmd = ['ec2', 'describe-instances']
+            cmd += ['--filters', f'Name=tag:elasticbeanstalk:environment-name,Values={eb_name}']
+            rr = aws_cli_for_ec2.run(cmd)
+            if rr['Reservations'] and rr['Reservations'][0]['Instances']:
+                in_use_ec2_ami_list.append(rr['Reservations'][0]['Instances'][0]['ImageId'])
 
     print_message('get All imagebuilder resouce version')
-    cmd = ['imagebuilder', 'list-images']
-    rr = aws_cli.run(cmd)
-    ami_arn_list = filter_imagebuilder_resource_arn_list(rr['imageVersionList'])
+
+    print_message('describe log groups imagebuilder resouce')
+    cmd = ['logs', 'describe-log-groups']
+    cmd += ['--log-group-name-prefix', '/aws/imagebuilder/gendo']
+    rr = aws_cli.run(cmd, ignore_error=True)
+    imagebuilder_cw_log_list = rr['logGroups']
 
     cmd = ['ec2', 'describe-images']
     cmd += ['--filters=Name=name,Values="Gendo*"']
     cmd += ['--owners', account_id]
     rr = aws_cli.run(cmd)
     ec2_gendo_img_list = rr['Images']
+
+    in_use_ami_timestamp_version = list()
+    for img in ec2_gendo_img_list:
+        if img['ImageId'] in in_use_ec2_ami_list:
+            continue
+        for tag in img['Tags']:
+            if tag['Key'] == 'Ec2ImageBuilderArn':
+                m = re.search('/gendo-recipe-(.+?)/', tag['Value'])
+                in_use_ami_timestamp_version.append(m.group(1))
+
+    cmd = ['imagebuilder', 'list-images']
+    rr = aws_cli.run(cmd)
+    ami_arn_list = filter_imagebuilder_resource_arn_list(rr['imageVersionList'])
 
     cmd = ['imagebuilder', 'list-image-pipelines']
     rr = aws_cli.run(cmd)
@@ -142,17 +181,25 @@ def run_terminate_image(name):
     rr = aws_cli.run(cmd)
     image_recipe_list = filter_imagebuilder_resource_arn_list(rr['imageRecipeSummaryList'])
 
-    cmd = ['logs', 'describe-log-groups']
-    cmd += ['--log-group-name-prefix', '/aws/imagebuilder/gendo']
-    rr = aws_cli.run(cmd, ignore_error=True)
-    imagebuilder_cw_log_list = rr['logGroups']
+    print_message('get component list')
+    cmd = ['imagebuilder', 'list-components']
+    rr = aws_cli.run(cmd)
+
+    component_list = filter_imagebuilder_resource_arn_list(rr['componentVersionList'])
+    timestamp_list = list()
+    for cc in component_list:
+        found = re.search('-component-(.+?)/', cc).group(1)
+        if found:
+            timestamp_list.append(found)
+    timestamp_list = set(timestamp_list)
 
     imagebuilder_resource['imagebuilder_cw_log_list'] = imagebuilder_cw_log_list
-    imagebuilder_resource['image_recipe_list'] = image_recipe_list
-    imagebuilder_resource['distribution_list'] = distribution_list
-    imagebuilder_resource['pipe_line_list'] = pipe_line_list
-    imagebuilder_resource['ec2_gendo_img_list'] = ec2_gendo_img_list
     imagebuilder_resource['ami_arn_list'] = ami_arn_list
+    imagebuilder_resource['ec2_gendo_img_list'] = ec2_gendo_img_list
+    imagebuilder_resource['pipe_line_list'] = pipe_line_list
+    imagebuilder_resource['distribution_list'] = distribution_list
+    imagebuilder_resource['infrastructure_list'] = infrastructure_list
+    imagebuilder_resource['image_recipe_list'] = image_recipe_list
     imagebuilder_resource['component_list'] = component_list
 
     normal_resource_version_list = list()
@@ -164,10 +211,33 @@ def run_terminate_image(name):
         else:
             abnormal_resource_version_list.append(tt)
 
+    print_message(f'normal resource version list : {normal_resource_version_list}')
+    print_message(f'abnormal resource version list : {abnormal_resource_version_list}')
+
     delete_version_list = list(abnormal_resource_version_list)
-    if len(normal_resource_version_list) > 2:
-        normal_resource_version_list.sort(key=int)
-        delete_version_list += normal_resource_version_list[2:]
+
+    tt = datetime.now() - timedelta(weeks=8)
+    timestamp_8_weeks_ago = time.mktime(tt.timetuple())
+
+    for vv in normal_resource_version_list:
+        if vv in in_use_ami_timestamp_version:
+            continue
+
+        if int(vv) < int(timestamp_8_weeks_ago):
+            delete_version_list.append(vv)
+
+    if not delete_version_list:
+        print_message('There are no versions to delete.')
+        return
+
+    print_message(f'delete version list : {delete_version_list}')
+
+    print_message('delete cloudwatch image builder logs')
+    for log_group in imagebuilder_cw_log_list:
+        if delete_version_list_any_cloudwatch_log(delete_version_list, log_group):
+            cmd = ['logs', 'delete-log-group']
+            cmd += ['--log-group-name', log_group['logGroupName']]
+            aws_cli.run(cmd, ignore_error=True)
 
     for ami_arn in ami_arn_list:
         if delete_version_list_any_imagebuilder_resource(delete_version_list, ami_arn):
@@ -240,13 +310,6 @@ def run_terminate_image(name):
                     cmd = ['imagebuilder', 'delete-component']
                     cmd += ['--component-build-version-arn', r['arn']]
                     aws_cli.run(cmd, ignore_error=True)
-
-    print_message('delete cloudwatch image builder logs')
-    for log_group in imagebuilder_cw_log_list:
-        if delete_version_list_any_cloudwatch_log(delete_version_list, log_group):
-            cmd = ['logs', 'delete-log-group']
-            cmd += ['--log-group-name', log_group['logGroupName']]
-            aws_cli.run(cmd, ignore_error=True)
 
 
 ################################################################################
